+++ lib/loggerator/middleware.rb
@@ -1,5 +1,5 @@
-require_relative './middleware/request_store'
-require_relative './middleware/request_id'
+require_relative 'middleware/request_store'
+require_relative 'middleware/request_id'
 
 module Loggerator
   module Middleware; end
